Safari-609.1.11.1/LayoutTests/http/tests/security/cross-frame-access-location-put-expected.txt


SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.


--------
Frame: '<!--frame1-->'
--------

----- tests for putting window.history and its properties -----

PASS: window.location.assign should be 'function assign() {    [native code]}' and is.
PASS: window.location.replace should be 'function replace() {    [native code]}' and is.
PASS: window.location.reload should be 'function reload() {    [native code]}' and is.
PASS: window.location.toString should be 'function toString() {    [native code]}' and is.
PASS: window.location.customAttribute should be 'customAttribute' and is.
